Moves `FeedCache` and `FeedLoaderCacheDecorator` to production.
