+++ FeedApplication/FeedApplication/FeedLoaderCacheDecorator.swift
@@ -1,0 +1,27 @@
+//
+//  FeedLoaderCacheDecorator.swift
+//  FeedApplication
+//
+//  Created by Omran Khoja on 3/13/22.
+//
+
+import FeedCoreModule
+
+public class FeedLoaderCacheDecorator: FeedLoader {
+    let decoratee: FeedLoader
+    let cache: FeedCache
+    
+    public init(decoratee: FeedLoader, cache: FeedCache) {
+        self.decoratee = decoratee
+        self.cache = cache
+    }
+    
+    public func load(completion: @escaping (FeedLoader.Result) -> Void) {
+        decoratee.load { [weak self] result in
+            if let feed = try? result.get() {
+                self?.cache.save(feed) { _ in }
+            }
+            completion(result)
+        }
+    }
+}
